include keycloak-enabled tests in travis build
