--- activiti-services/activiti-services-identity-keycloak/src/test/java/org/activiti/services/identity/keycloak/Application.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2017 Alfresco, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.activiti.services.identity.keycloak;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.PropertySource;
-
-@SpringBootApplication
-@ComponentScan("org.activiti")
-public class Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class,
-                              args);
-    }
-}
