--- activiti-services/activiti-services-identity-keycloak/src/test/resources/application-test.properties
@@ -1,2 +1,0 @@
-keycloaktestuser=testuser
-keycloaktestpassword=password
