--- activiti-services/activiti-services-identity-keycloak/src/test/resources/processes/SimpleProcess.bpmn20.xml
@@ -1,57 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<definitions xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:activiti="http://activiti.org/bpmn" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:omgdc="http://www.omg.org/spec/DD/20100524/DC" xmlns:omgdi="http://www.omg.org/spec/DD/20100524/DI" typeLanguage="http://www.w3.org/2001/XMLSchema" expressionLanguage="http://www.w3.org/1999/XPath" targetNamespace="http://www.activiti.org/processdef" xmlns:modeler="http://activiti.com/modeler" modeler:version="1.0en" modeler:exportDateTime="20170630121556884" modeler:modelId="31032" modeler:modelVersion="1" modeler:modelLastUpdated="1498821282568">
-  <process id="SimpleProcess" name="SimpleProcess" isExecutable="true">
-    <startEvent id="startEvent1">
-      <extensionElements>
-        <modeler:editor-resource-id><![CDATA[startEvent1]]></modeler:editor-resource-id>
-      </extensionElements>
-    </startEvent>
-    <userTask id="sid-CDFE7219-4627-43E9-8CA8-866CC38EBA94" name="Perform action" activiti:candidateGroups="hr">
-      <extensionElements>
-        <modeler:allow-send-email><![CDATA[true]]></modeler:allow-send-email>
-        <modeler:initiator-can-complete><![CDATA[false]]></modeler:initiator-can-complete>
-        <modeler:editor-resource-id><![CDATA[sid-CDFE7219-4627-43E9-8CA8-866CC38EBA94]]></modeler:editor-resource-id>
-      </extensionElements>
-    </userTask>
-    <sequenceFlow id="sid-68945AF1-396F-4B8A-B836-FC318F62313F" sourceRef="startEvent1" targetRef="sid-CDFE7219-4627-43E9-8CA8-866CC38EBA94">
-      <extensionElements>
-        <modeler:editor-resource-id><![CDATA[sid-68945AF1-396F-4B8A-B836-FC318F62313F]]></modeler:editor-resource-id>
-      </extensionElements>
-    </sequenceFlow>
-    <endEvent id="sid-9F938772-895F-436D-A7D1-051B7A69D676">
-      <extensionElements>
-        <modeler:editor-resource-id><![CDATA[sid-9F938772-895F-436D-A7D1-051B7A69D676]]></modeler:editor-resource-id>
-      </extensionElements>
-    </endEvent>
-    <sequenceFlow id="sid-3C0516C4-1BF7-48F8-88CF-04C41810FE00" sourceRef="sid-CDFE7219-4627-43E9-8CA8-866CC38EBA94" targetRef="sid-9F938772-895F-436D-A7D1-051B7A69D676">
-      <extensionElements>
-        <modeler:editor-resource-id><![CDATA[sid-3C0516C4-1BF7-48F8-88CF-04C41810FE00]]></modeler:editor-resource-id>
-      </extensionElements>
-    </sequenceFlow>
-  </process>
-  <bpmndi:BPMNDiagram id="BPMNDiagram_SimpleProcess">
-    <bpmndi:BPMNPlane bpmnElement="SimpleProcess" id="BPMNPlane_SimpleProcess">
-      <bpmndi:BPMNShape bpmnElement="startEvent1" id="BPMNShape_startEvent1">
-        <omgdc:Bounds height="30.0" width="30.0" x="100.0" y="163.0"/>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape bpmnElement="sid-CDFE7219-4627-43E9-8CA8-866CC38EBA94" id="BPMNShape_sid-CDFE7219-4627-43E9-8CA8-866CC38EBA94">
-        <omgdc:Bounds height="80.0" width="100.0" x="195.0" y="135.0"/>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape bpmnElement="sid-9F938772-895F-436D-A7D1-051B7A69D676" id="BPMNShape_sid-9F938772-895F-436D-A7D1-051B7A69D676">
-        <omgdc:Bounds height="28.0" width="28.0" x="360.0" y="164.0"/>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNEdge bpmnElement="sid-68945AF1-396F-4B8A-B836-FC318F62313F" id="BPMNEdge_sid-68945AF1-396F-4B8A-B836-FC318F62313F">
-        <omgdi:waypoint x="130.0" y="178.0"/>
-        <omgdi:waypoint x="162.5" y="178.0"/>
-        <omgdi:waypoint x="162.5" y="175.0"/>
-        <omgdi:waypoint x="195.0" y="175.0"/>
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNEdge bpmnElement="sid-3C0516C4-1BF7-48F8-88CF-04C41810FE00" id="BPMNEdge_sid-3C0516C4-1BF7-48F8-88CF-04C41810FE00">
-        <omgdi:waypoint x="295.0" y="175.0"/>
-        <omgdi:waypoint x="327.5" y="175.0"/>
-        <omgdi:waypoint x="327.5" y="178.0"/>
-        <omgdi:waypoint x="360.0" y="178.0"/>
-      </bpmndi:BPMNEdge>
-    </bpmndi:BPMNPlane>
-  </bpmndi:BPMNDiagram>
-</definitions>
