--- activiti-spring-boot/spring-boot-samples/spring-boot-sample-hal-rest-api/src/test/resources/application-test.properties
@@ -1,2 +1,0 @@
-keycloaktestuser=testuser
-keycloaktestpassword=password
