--- activiti-spring-boot/spring-boot-starters/activiti-spring-boot-starter/src/test/java/org/activiti/runtime/ProcessInstanceIT.java
@@ -1,168 +1,0 @@
-/*
- * Copyright 2017 Alfresco, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.activiti.runtime;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import org.activiti.definition.ProcessDefinitionIT;
-import org.activiti.services.core.model.ProcessDefinition;
-import org.activiti.services.core.model.ProcessInstance;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.hateoas.PagedResources;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.junit4.SpringRunner;
-
-import static org.activiti.runtime.ProcessInstanceRestTemplate.PROCESS_INSTANCES_RELATIVE_URL;
-import static org.assertj.core.api.Assertions.*;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class ProcessInstanceIT {
-
-    private static final String SIMPLE_PROCESS = "SimpleProcess";
-
-    @Autowired
-    private TestRestTemplate restTemplate;
-
-    @Autowired
-    private ProcessInstanceRestTemplate processInstanceRestTemplate;
-
-    private Map<String, String> processDefinitionIds = new HashMap<>();
-
-    @Before
-    public void setUp() {
-        ResponseEntity<PagedResources<ProcessDefinition>> processDefinitions = getProcessDefinitions();
-        assertThat(processDefinitions.getBody().getContent()).hasSize(3);
-        for (ProcessDefinition pd : processDefinitions.getBody().getContent()) {
-            processDefinitionIds.put(pd.getName(),
-                                     pd.getId());
-        }
-    }
-
-    @Test
-    public void shouldStartProcess() throws Exception {
-        //when
-        ResponseEntity<ProcessInstance> entity = processInstanceRestTemplate.startProcess(processDefinitionIds.get(SIMPLE_PROCESS));
-
-        //then
-        assertThat(entity).isNotNull();
-        ProcessInstance returnedProcInst = entity.getBody();
-        assertThat(returnedProcInst).isNotNull();
-        assertThat(returnedProcInst.getId()).isNotNull();
-        assertThat(returnedProcInst.getProcessDefinitionId()).contains("SimpleProcess:");
-    }
-
-    @Test
-    public void shouldRetrieveProcessInstanceById() throws Exception {
-        //given
-        ResponseEntity<ProcessInstance> startedProcessEntity = processInstanceRestTemplate.startProcess(processDefinitionIds.get(SIMPLE_PROCESS));
-
-        //when
-        ResponseEntity<ProcessInstance> retrievedEntity = restTemplate.exchange(
-                PROCESS_INSTANCES_RELATIVE_URL + startedProcessEntity.getBody().getId(),
-                HttpMethod.GET,
-                null,
-                new ParameterizedTypeReference<ProcessInstance>() {
-                });
-
-        //then
-        assertThat(retrievedEntity.getBody()).isNotNull();
-    }
-
-    @Test
-    public void shouldRetrieveListOfProcessInstances() throws Exception {
-        //given
-        processInstanceRestTemplate.startProcess(processDefinitionIds.get(SIMPLE_PROCESS));
-        processInstanceRestTemplate.startProcess(processDefinitionIds.get(SIMPLE_PROCESS));
-        processInstanceRestTemplate.startProcess(processDefinitionIds.get(SIMPLE_PROCESS));
-
-        //when
-        ResponseEntity<PagedResources<ProcessInstance>> processInstancesPage = restTemplate.exchange(PROCESS_INSTANCES_RELATIVE_URL + "?page={page}&size={size}",
-                                                                                                     HttpMethod.GET,
-                                                                                                     null,
-                                                                                                     new ParameterizedTypeReference<PagedResources<ProcessInstance>>() {
-                                                                                                     },
-                                                                                                     "0",
-                                                                                                     "2");
-
-        //then
-        assertThat(processInstancesPage).isNotNull();
-        assertThat(processInstancesPage.getBody().getContent()).hasSize(2);
-        assertThat(processInstancesPage.getBody().getMetadata().getTotalPages()).isGreaterThanOrEqualTo(3);
-    }
-
-    @Test
-    public void suspendShouldPutProcessInstanceInSuspendedState() throws Exception {
-        //given
-        ResponseEntity<ProcessInstance> startProcessEntity = processInstanceRestTemplate.startProcess(processDefinitionIds.get(SIMPLE_PROCESS));
-
-        //when
-        ResponseEntity<Void> responseEntity = executeRequestSuspendProcess(startProcessEntity);
-
-        //then
-        assertThat(responseEntity.getStatusCode()).isEqualTo(HttpStatus.OK);
-        ResponseEntity<ProcessInstance> processInstanceEntity = processInstanceRestTemplate.getProcessInstance(startProcessEntity);
-        assertThat(processInstanceEntity.getBody().getStatus()).isEqualTo(ProcessInstance.ProcessInstanceStatus.SUSPENDED.name());
-    }
-
-    private ResponseEntity<Void> executeRequestSuspendProcess(ResponseEntity<ProcessInstance> processInstanceEntity) {
-        ResponseEntity<Void> responseEntity = restTemplate.exchange(PROCESS_INSTANCES_RELATIVE_URL + processInstanceEntity.getBody().getId() + "/suspend",
-                                                                    HttpMethod.POST,
-                                                                    null,
-                                                                    new ParameterizedTypeReference<Void>() {
-                                                                    });
-        assertThat(responseEntity.getStatusCode()).isEqualTo(HttpStatus.OK);
-        return responseEntity;
-    }
-
-    @Test
-    public void activateShouldPutASuspendedProcessInstanceBackToActiveState() throws Exception {
-        //given
-        ResponseEntity<ProcessInstance> startProcessEntity = processInstanceRestTemplate.startProcess(processDefinitionIds.get(SIMPLE_PROCESS));
-        executeRequestSuspendProcess(startProcessEntity);
-
-        //when
-        ResponseEntity<Void> responseEntity = restTemplate.exchange(PROCESS_INSTANCES_RELATIVE_URL + startProcessEntity.getBody().getId() + "/activate",
-                                                                    HttpMethod.POST,
-                                                                    null,
-                                                                    new ParameterizedTypeReference<Void>() {
-                                                                    });
-
-        //then
-        assertThat(responseEntity.getStatusCode()).isEqualTo(HttpStatus.OK);
-        ResponseEntity<ProcessInstance> processInstanceEntity = processInstanceRestTemplate.getProcessInstance(startProcessEntity);
-        assertThat(processInstanceEntity.getBody().getStatus()).isEqualTo(ProcessInstance.ProcessInstanceStatus.RUNNING.name());
-    }
-
-    private ResponseEntity<PagedResources<ProcessDefinition>> getProcessDefinitions() {
-        ParameterizedTypeReference<PagedResources<ProcessDefinition>> responseType = new ParameterizedTypeReference<PagedResources<ProcessDefinition>>() {
-        };
-        return restTemplate.exchange(ProcessDefinitionIT.PROCESS_DEFINITIONS_URL,
-                                     HttpMethod.GET,
-                                     null,
-                                     responseType);
-    }
-}
