+++ activiti-spring-boot/spring-boot-starters/activiti-spring-boot-starter/src/test/java/org/activiti/runtime/ProcessVariablesKeycloakIT.java
@@ -19,7 +19,9 @@ package org.activiti.runtime;
 import java.util.HashMap;
 import java.util.Map;
 
-import org.activiti.definition.ProcessDefinitionIT;
+import org.activiti.definition.ProcessDefinitionKeycloakIT;
+import org.activiti.keycloak.KeycloakEnabledBaseTestIT;
+import org.activiti.keycloak.ProcessInstanceKeycloakRestTemplate;
 import org.activiti.services.core.model.ProcessDefinition;
 import org.activiti.services.core.model.ProcessInstance;
 import org.junit.Before;
@@ -33,26 +35,30 @@ import org.springframework.hateoas.PagedResources;
 import org.springframework.hateoas.Resource;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
+import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
+import static org.activiti.keycloak.ProcessInstanceKeycloakRestTemplate.PROCESS_INSTANCES_RELATIVE_URL;
 
 import static org.assertj.core.api.Assertions.*;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class ProcessVariablesIT {
+@TestPropertySource("classpath:application-test.properties")
+public class ProcessVariablesKeycloakIT extends KeycloakEnabledBaseTestIT {
 
     @Autowired
     private TestRestTemplate restTemplate;
 
     @Autowired
-    private ProcessInstanceRestTemplate processInstanceRestTemplate;
+    private ProcessInstanceKeycloakRestTemplate processInstanceRestTemplate;
 
     private Map<String, String> processDefinitionIds = new HashMap<>();
 
     private static final String SIMPLE_PROCESS_WITH_VARIABLES = "ProcessWithVariables";
 
     @Before
-    public void setUp() {
+    public void setUp() throws Exception {
+        super.setUp();
         ResponseEntity<PagedResources<ProcessDefinition>> processDefinitions = getProcessDefinitions();
         assertThat(processDefinitions.getBody().getContent()).hasSize(3);
         for (ProcessDefinition pd : processDefinitions.getBody().getContent()) {
@@ -72,12 +78,12 @@ public class ProcessVariablesIT {
         variables.put("age",
                       15);
         ResponseEntity<ProcessInstance> startResponse = processInstanceRestTemplate.startProcess(processDefinitionIds.get(SIMPLE_PROCESS_WITH_VARIABLES),
-                                                                                                 variables);
+                                                                                                 variables,accessToken);
 
         //when
-        ResponseEntity<Resource<Map<String, Object>>> variablesResponse = restTemplate.exchange(ProcessInstanceRestTemplate.PROCESS_INSTANCES_RELATIVE_URL + startResponse.getBody().getId() + "/variables",
+        ResponseEntity<Resource<Map<String, Object>>> variablesResponse = restTemplate.exchange(PROCESS_INSTANCES_RELATIVE_URL + startResponse.getBody().getId() + "/variables",
                                                                                                 HttpMethod.GET,
-                                                                                                null,
+                                                                                                getRequestEntityWithHeaders(),
                                                                                                 new ParameterizedTypeReference<Resource<Map<String, Object>>>() {
                                                                                                 });
 
@@ -95,9 +101,9 @@ public class ProcessVariablesIT {
     private ResponseEntity<PagedResources<ProcessDefinition>> getProcessDefinitions() {
         ParameterizedTypeReference<PagedResources<ProcessDefinition>> responseType = new ParameterizedTypeReference<PagedResources<ProcessDefinition>>() {
         };
-        return restTemplate.exchange(ProcessDefinitionIT.PROCESS_DEFINITIONS_URL,
+        return restTemplate.exchange(ProcessDefinitionKeycloakIT.PROCESS_DEFINITIONS_URL,
                                      HttpMethod.GET,
-                                     null,
+                                     getRequestEntityWithHeaders(),
                                      responseType);
     }
 }
