+++ docker/README.md
@@ -1,0 +1,5 @@
+# Docker containers for Activiti
+
+Dockerfile is for keycloak. It uses springboot-realm.json as a config file. See sample project.
+
+docker-compose.yml is for rabbitmq.
