+++ pom.xml
@@ -631,7 +631,8 @@
           <excludes>
             <exclude>**/*TestCase.java</exclude>
             <exclude>**/*IT.java</exclude>
-            <exclude>**/*Keycloak*.java</exclude>
+            <!--no longer exclude keycloak as travis should be running it
+            <exclude>**/*Keycloak*.java</exclude> -->
           </excludes>
           <runOrder>alphabetical</runOrder>
         </configuration>
@@ -869,11 +870,12 @@
           <groupId>org.apache.maven.plugins</groupId>
           <artifactId>maven-failsafe-plugin</artifactId>
           <version>${maven-failsafe-plugin.version}</version>
+          <!--no longer exclude keycloak as travis should be running it
           <configuration>
             <excludes>
               <exclude>**/*Keycloak*.java</exclude>
             </excludes>
-          </configuration>
+          </configuration> -->
           <executions>
             <execution>
               <id>integration-test</id>
